Moved some things around and more rubocop fixes.
