+++ .rubocop.yml
@@ -67,8 +67,10 @@ RSpec/PendingWithoutReason:
 
 RSpec/EmptyExampleGroup:
   Exclude:
-    - 'spec/requests/api/v1/organizations_spec.rb'
+    - 'spec/requests/api/v1/api_docs/organizations_spec.rb'
+    - 'spec/requests/api/v1/api_docs/publications_spec.rb'
+    - 'spec/requests/api/v1/api_docs/users_spec.rb'
 
 RSpec/VariableName:
   Exclude:
-    - 'spec/requests/api/v1/organizations_spec.rb'
+    - 'spec/requests/api/v1/api_docs/organizations_spec.rb'
